pyproject.toml
[tool.poetry]
name = "ascv-genai"
version = "0.1.0"
description = "Autonomous Safety Constraint Validation - Generative AI Pipeline"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "ascv_genai", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
python-multipart = "^0.0.6"
langchain = "^0.0.345"
openai = "^1.3.7"
chromadb = "^0.4.18"
pandas = "^2.1.3"
protobuf = "^4.25.1"
requests = "^2.31.0"
celery = "^5.3.6"
redis = "^5.0.1"
python-decouple = "^3.8"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
httpx = "^0.25.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
